Ask for confirmation before deleting/saving in navview. Fix #30.

+++ static/visualizer/src/header/components/navview.js
@@ -74,25 +74,29 @@ define(['jquery', 'components/superagent/superagent', 'src/header/components/def
             var that = this;
             var dir = this.getDir(this.activeNode.data.path);
             var name = this.activeNode.title;
-            console.log('dir to save', dir);
-            var req = $.ajax({
-                url: '/navview/save',
-                data: {
-                    dir: dir,
-                    name: name,
-                    content: Versioning.getViewJSON("\t")
-                },
-                type: 'POST',
-                dataType: 'json'
+
+            confirm('You are about to save the current view to: ' + this.activeNode.data.path + '<br/>This operation will erase the previous content of this file and cannot be undone.').then(function(ok) {
+                if(!ok) return;
+                var req = $.ajax({
+                    url: '/navview/save',
+                    data: {
+                        dir: dir,
+                        name: name,
+                        content: Versioning.getViewJSON("\t")
+                    },
+                    type: 'POST',
+                    dataType: 'json'
+                });
+
+                req.done(function() {
+                    that.log('success-log', 'Successfully saved view');
+                    that.reloadActiveNode();
+                });
+                req.fail(function() {
+                    that.log('error-log', 'Failed to save view');
+                });
             });
 
-            req.done(function() {
-                that.log('success-log', 'Successfully saved view');
-                that.reloadActiveNode();
-            });
-            req.fail(function() {
-                that.log('error-log', 'Failed to save view');
-            });
         },
 
         mkdir: function() {
@@ -132,25 +136,29 @@ define(['jquery', 'components/superagent/superagent', 'src/header/components/def
 
             var dir = this.getDir(node.data.path);
             var name = node.title;
-            console.log(dir, name);
-            var req = $.ajax({
-                url: '/navview/file',
-                type: 'DELETE',
-                data: {
-                    dir: dir,
-                    name: name
-                },
-                dataType: 'json'
+
+            confirm('<p>You are about to remove the file: <br/>' + node.data.path + '</p>Do you really want to do this? You cannot undo this operation').then(function(ok) {
+                if(!ok) return;
+                var req = $.ajax({
+                    url: '/navview/file',
+                    type: 'DELETE',
+                    data: {
+                        dir: dir,
+                        name: name
+                    },
+                    dataType: 'json'
+                });
+
+                req.done(function() {
+                    node.remove();
+                    that.log('success-log', 'Successfully removed file');
+                });
+
+                req.fail(function() {
+                    that.log('error-log', 'Failed remove file');
+                });
             });
 
-            req.done(function() {
-                node.remove();
-                that.log('success-log', 'Successfully removed file');
-            });
-
-            req.fail(function() {
-                that.log('error-log', 'Failed remove file');
-            });
         },
 
         removeDir: function(node) {
@@ -159,23 +167,29 @@ define(['jquery', 'components/superagent/superagent', 'src/header/components/def
                 return this.log('error-log', 'Failed remove directory');
             }
 
-            var req = $.ajax({
-                url: '/navview/dir',
-                type: 'DELETE',
-                data: {
-                    dir: node.data.path
-                },
-                dataType: 'json'
+            confirm('<p>You are about to remove the directory: <br/>' + node.data.path + '</p>Do you really want to do this? You cannot undo this operation').then(function(ok) {
+                if(!ok) return;
+                var req = $.ajax({
+                    url: '/navview/dir',
+                    type: 'DELETE',
+                    data: {
+                        dir: node.data.path
+                    },
+                    dataType: 'json'
+                });
+
+                req.done(function() {
+                    node.remove();
+                    that.log('success-log', 'Successfully removed directory');
+                });
+
+                req.fail(function() {
+                    that.log('error-log', 'Failed remove directory');
+                });
             });
 
-            req.done(function() {
-                node.remove();
-                that.log('success-log', 'Successfully removed directory');
-            });
 
-            req.fail(function() {
-                that.log('error-log', 'Failed remove directory');
-            });
+
         },
 
         rename: function() {
@@ -456,4 +470,33 @@ function fancyTreeDirStructure(list) {
     });
     console.log('XXXX',  x);
     return x;
+}
+
+function confirm(message) {
+    return new Promise(function(resolve){
+        var $dialog = $('#ci-dialog');
+        if($dialog.length === 0) {
+            $dialog = $('<div/>').css('id', 'ci-dialog');
+            $('body').append($dialog);
+        }
+        $dialog.html(message);
+        $dialog.dialog({
+            modal: true,
+            buttons: {
+                Cancel: function() {
+                    $(this).dialog('close');
+                },
+                Ok: function() {
+                    console.log('ok...');
+                    resolve(true);
+                    $(this).dialog('close');
+                }
+            },
+            close: function() {
+                return resolve(false);
+            },
+            width: 400
+        });
+    });
+
 }
